+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+#### Fixed
+- `?for_complete=true` added in stop api call
+
+#### Added
+- Support for sending start scan request
+
 ## [0.0.4] - 2023-08-18
 - Docker image tag: `0.0.4-68c1832`
 #### Fixed